tests/wpt/web-platform-tests/offscreen-canvas/shadows/2d.shadow.offset.negativeY.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.shadow.offset.negativeY</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.shadow.offset.negativeY</h1>
<p class="desc">Shadows can be offset with negative y</p>


<script>
var t = async_test("Shadows can be offset with negative y");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.fillStyle = '#0f0';
ctx.shadowColor = '#0f0';
ctx.shadowOffsetY = -25;
ctx.fillRect(0, 25, 100, 25);
_assertPixel(offscreenCanvas, 50,12, 0,255,0,255, "50,12", "0,255,0,255");
_assertPixel(offscreenCanvas, 50,37, 0,255,0,255, "50,37", "0,255,0,255");

t.done();

});
</script>
